dealing with circular references
