+++ microservices/product-composite-service/src/main/java/com/yavlash/microservices/composite/product/config/ApplicationConfiguration.java
@@ -1,6 +1,6 @@
 package com.yavlash.microservices.composite.product.config;
 
-import com.yavlash.microservices.composite.product.services.ProductCompositeIntegration;
+import com.yavlash.microservices.composite.product.services.HealthService;
 import io.swagger.v3.oas.models.ExternalDocumentation;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Contact;
@@ -14,8 +14,8 @@ import org.springframework.boot.actuate.health.ReactiveHealthIndicator;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -71,14 +71,14 @@ public class ApplicationConfiguration {
     }
 
     @Autowired
-    ProductCompositeIntegration integration;
+    HealthService healthService;
 
     @Bean
     ReactiveHealthContributor coreServices() {
         final Map<String, ReactiveHealthIndicator> registry = new LinkedHashMap<>();
-        registry.put("product", () -> integration.getProductHealth());
-        registry.put("recommendation", () -> integration.getRecommendationHealth());
-        registry.put("review", () -> integration.getReviewHealth());
+        registry.put("product", () -> healthService.getProductHealth());
+        registry.put("recommendation", () -> healthService.getRecommendationHealth());
+        registry.put("review", () -> healthService.getReviewHealth());
         return CompositeReactiveHealthContributor.fromMap(registry);
     }
