+++ microservices/product-composite-service/src/main/java/com/yavlash/microservices/composite/product/services/HealthService.java
@@ -1,0 +1,48 @@
+package com.yavlash.microservices.composite.product.services;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.actuate.health.Health;
+import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.publisher.Mono;
+
+import static java.util.logging.Level.FINE;
+
+@Component
+public class HealthService {
+    private static final Logger LOG = LoggerFactory.getLogger(HealthService.class);
+    private static final String PRODUCT_SERVICE_URL = "http://product";
+    private static final String RECOMMENDATION_SERVICE_URL = "http://recommendation";
+    private static final String REVIEW_SERVICE_URL = "http://review";
+    private final WebClient webClient;
+
+    @Autowired
+    public HealthService(
+            WebClient.Builder webClientBuilder
+    ) {
+        this.webClient = webClientBuilder.build();
+    }
+
+    public Mono<Health> getProductHealth() {
+        return getHealth(PRODUCT_SERVICE_URL);
+    }
+
+    public Mono<Health> getRecommendationHealth() {
+        return getHealth(RECOMMENDATION_SERVICE_URL);
+    }
+
+    public Mono<Health> getReviewHealth() {
+        return getHealth(REVIEW_SERVICE_URL);
+    }
+
+    private Mono<Health> getHealth(String url) {
+        url += "/actuator/health";
+        LOG.debug("Will call the Health API on URL: {}", url);
+        return webClient.get().uri(url).retrieve().bodyToMono(String.class)
+                .map(s -> new Health.Builder().up().build())
+                .onErrorResume(ex -> Mono.just(new Health.Builder().down(ex).build()))
+                .log(LOG.getName(), FINE);
+    }
+}
